requirements.txt
functions-framework==3.*
loguru
flask
numpy
openai
openai-agents
codename
google-cloud-firestore